+++ classes/Requires.php
@@ -17,6 +17,7 @@
 	
  	class Requires {
  		static function autoload() {
+ 			require_once(dirname(__DIR__) . "/classes/Prerequisites.php");
  			require_once(dirname(__DIR__) . "/classes/Error.php");
  			require_once(dirname(__DIR__) . "/app.config.php");